Made the RecordExistsQueryFormatter and the CovariateManager common. Also removed the interface for AgeGenderYearManager, as there was only going to be a single implementation.
