+++ rifServices/src/main/resources/RIFLogger.properties
@@ -113,7 +113,7 @@ rifServices.restfulWebServices.ms.MSSQLRIFStudySubmissionWebServiceResource=true
 rifServices.dataStorageLayer.ms.MSSQLResultsQueryManager=true
 rifServices.dataStorageLayer.ms.MSSQLRIFContextManager=true
 rifServices.dataStorageLayer.ms.MSSQLConnectionManager=true
-rifServices.dataStorageLayer.common.CommonAgeGenderYearManager=true
+rifServices.dataStorageLayer.common.AgeGenderYearManager=true
 rifServices.dataStorageLayer.ms.MSSQLCreateStudySubmissionStep=true
 rifServices.dataStorageLayer.ms.MSSQLStudyStateManager=true
 rifServices.dataStorageLayer.ms.MSSQLGenerateResultsSubmissionStep=true
